Move `TestCustomisableModRuleset` out of `TestSceneModSettings`
