+++ osu.Game.Tests/Visual/Gameplay/TestSceneSpectatorPlayback.cs
@@ -27,8 +27,8 @@ using osu.Game.Rulesets.Replays.Types;
 using osu.Game.Rulesets.UI;
 using osu.Game.Scoring;
 using osu.Game.Screens.Play;
+using osu.Game.Tests.Mods;
 using osu.Game.Tests.Visual.Spectator;
-using osu.Game.Tests.Visual.UserInterface;
 using osuTK;
 using osuTK.Graphics;
 
@@ -77,7 +77,7 @@ namespace osu.Game.Tests.Visual.Gameplay
                             {
                                 new Drawable[]
                                 {
-                                    recordingManager = new TestRulesetInputManager(TestSceneModSettings.CreateTestRulesetInfo(), 0, SimultaneousBindingMode.Unique)
+                                    recordingManager = new TestRulesetInputManager(TestCustomisableModRuleset.CreateTestRulesetInfo(), 0, SimultaneousBindingMode.Unique)
                                     {
                                         Recorder = recorder = new TestReplayRecorder
                                         {
@@ -107,7 +107,7 @@ namespace osu.Game.Tests.Visual.Gameplay
                                 },
                                 new Drawable[]
                                 {
-                                    playbackManager = new TestRulesetInputManager(TestSceneModSettings.CreateTestRulesetInfo(), 0, SimultaneousBindingMode.Unique)
+                                    playbackManager = new TestRulesetInputManager(TestCustomisableModRuleset.CreateTestRulesetInfo(), 0, SimultaneousBindingMode.Unique)
                                     {
                                         Clock = new FramedClock(manualClock),
                                         ReplayInputHandler = replayHandler = new TestFramedReplayInputHandler(replay)
